requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
ultralytics>=8.2.0
pillow>=10.0.0
httpx>=0.27.0
python-multipart>=0.0.9
